feat(initial): Initial version of HiMama downloader
This is a very simple tool to download all the photos from your HiMama
account and attach the metadata to those images. I'd rather not attach
the metadata with `exiftool`, but it was the only way that I could
figure out how to attach EXIF, IPTC, and XMP metadata in such a way that
it gets picked up by Apple Photos.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>